fix WrapPcap Ustack.LinkEndpoint will close the stack, and link drop
